+++ packages/base-ui/src/components/dialog/search.tsx
@@ -17,16 +17,21 @@ import {
 import { I18nLabel, useI18n } from '@/contexts/i18n';
 import { cn } from '@/utils/cn';
 import { Dialog } from '@base-ui/react/dialog';
-import type { HighlightedText, ReactSortedResult as BaseResultType } from '@hanzo/docs-core/search';
+import type { HighlightedText, ReactSortedResult } from '@hanzo/docs-core/search';
 import { cva } from 'class-variance-authority';
 import { useRouter } from '@hanzo/docs-core/framework';
 import type { SharedProps } from '@/contexts/search';
 import { useOnChange } from '@hanzo/docs-core/utils/use-on-change';
 import scrollIntoView from 'scroll-into-view-if-needed';
 import { buttonVariants } from '@/components/ui/button';
+import { createMarkdownRenderer } from '@hanzo/docs-core/content/md';
+import rehypeRaw from 'rehype-raw';
+import { visit } from 'unist-util-visit';
+import type { Transformer } from 'unified';
+import type { Root } from 'hast';
 
 export type SearchItemType =
-  | (BaseResultType & {
+  | (ReactSortedResult & {
       external?: boolean;
     })
   | {
@@ -48,7 +53,7 @@ export interface SearchDialogProps extends SharedProps {
   children: ReactNode;
 }
 
-const Context = createContext<{
+const RootContext = createContext<{
   open: boolean;
   onOpenChange: (open: boolean) => void;
   search: string;
@@ -68,6 +73,97 @@ const TagsListContext = createContext<{
   allowClear: boolean;
 } | null>(null);
 
+const PreContext = createContext(false);
+
+const mdRenderer = createMarkdownRenderer({
+  remarkRehypeOptions: {
+    allowDangerousHtml: true,
+  },
+  rehypePlugins: [rehypeRaw, rehypeCustomElements],
+});
+
+const mdComponents = {
+  mark(props: ComponentProps<'mark'>) {
+    return <span {...props} className="text-fd-primary underline" />;
+  },
+  a: 'span',
+  p(props: ComponentProps<'p'>) {
+    return <p {...props} className="min-w-0" />;
+  },
+  strong(props: ComponentProps<'strong'>) {
+    return <strong {...props} className="text-fd-accent-foreground font-medium" />;
+  },
+  code(props: ComponentProps<'pre'>) {
+    // eslint-disable-next-line react-hooks/rules-of-hooks -- this is a component
+    const inPre = use(PreContext);
+    if (inPre)
+      return (
+        <code
+          {...props}
+          className="mask-[linear-gradient(to_bottom,white,white_30px,transparent_80px)]"
+        />
+      );
+
+    return (
+      <code
+        {...props}
+        className="border rounded-md px-px bg-fd-secondary text-fd-secondary-foreground"
+      />
+    );
+  },
+  custom({
+    _tagName = 'fragment',
+    children,
+    ...rest
+  }: Record<string, unknown> & { _tagName: string; children: ReactNode }) {
+    return (
+      <span className="inline-flex max-w-full items-center border p-0.5 rounded-md bg-fd-card text-fd-card-foreground divide-x divide-fd-border">
+        <code className="rounded-sm px-0.5 me-1 bg-fd-primary font-medium text-xs text-fd-primary-foreground border-none">
+          {_tagName}
+        </code>
+        {Object.entries(rest).map(([k, v]) => {
+          if (typeof v !== 'string') return;
+
+          return (
+            <code key={k} className="truncate text-xs text-fd-muted-foreground px-1">
+              <span className="text-fd-card-foreground">{k}: </span>
+              {v}
+            </code>
+          );
+        })}
+        {children && <span className="ps-1">{children}</span>}
+      </span>
+    );
+  },
+  pre(props: ComponentProps<'pre'>) {
+    return (
+      <pre
+        {...props}
+        className={cn(
+          'flex flex-col border rounded-md my-0.5 p-2 bg-fd-secondary text-fd-secondary-foreground max-h-20 overflow-hidden',
+          props.className,
+        )}
+      >
+        <PreContext value={true}>{props.children}</PreContext>
+      </pre>
+    );
+  },
+};
+
+function rehypeCustomElements(): Transformer<Root, Root> {
+  return (tree) => {
+    visit(tree, (node) => {
+      if (
+        node.type === 'element' &&
+        document.createElement(node.tagName) instanceof HTMLUnknownElement
+      ) {
+        node.properties._tagName = node.tagName;
+        node.tagName = 'custom';
+      }
+    });
+  };
+}
+
 export function SearchDialog({
   open,
   onOpenChange,
@@ -78,7 +174,11 @@ export function SearchDialog({
   children,
 }: SearchDialogProps) {
   const router = useRouter();
-  const onSelect = useEffectEvent((item: SearchItemType) => {
+  const onOpenChangeCallback = useRef(onOpenChange);
+  onOpenChangeCallback.current = onOpenChange;
+  const onSearchChangeCallback = useRef(onSearchChange);
+  onSearchChangeCallback.current = onSearchChange;
+  const onSelect = (item: SearchItemType) => {
     if (item.type === 'action') {
       item.onSelect();
     } else if (item.external) {
@@ -89,26 +189,27 @@ export function SearchDialog({
 
     onOpenChange(false);
     onSelectProp?.(item);
-  });
+  };
+  const onSelectCallback = useRef(onSelect);
+  onSelectCallback.current = onSelect;
 
   return (
     <Dialog.Root open={open} onOpenChange={onOpenChange}>
-      <Context.Provider
+      <RootContext
         value={useMemo(
           () => ({
             open,
-            onOpenChange,
             search,
-            onSearchChange,
-            // eslint-disable-next-line react-hooks/rules-of-hooks -- used in child components
-            onSelect,
             isLoading,
+            onOpenChange: (v) => onOpenChangeCallback.current(v),
+            onSearchChange: (v) => onSearchChangeCallback.current(v),
+            onSelect: (v) => onSelectCallback.current(v),
           }),
-          [isLoading, onOpenChange, onSearchChange, open, search],
+          [isLoading, open, search],
         )}
       >
         {children}
-      </Context.Provider>
+      </RootContext>
     </Dialog.Root>
   );
 }
@@ -171,7 +272,7 @@ export function SearchDialogOverlay({
       {...props}
       className={(s) =>
         cn(
-          'fixed inset-0 z-50 backdrop-blur-xs bg-fd-overlay data-[open]:animate-fd-fade-in data-[closed]:animate-fd-fade-out',
+          'fixed inset-0 z-50 backdrop-blur-xs bg-fd-overlay data-open:animate-fd-fade-in data-closed:animate-fd-fade-out',
           typeof className === 'function' ? className(s) : className,
         )
       }
@@ -189,11 +290,12 @@ export function SearchDialogContent({
   return (
     <Dialog.Portal>
       <Dialog.Popup
+        id="fd-search-dialog-content"
         aria-describedby={undefined}
         {...props}
         className={(s) =>
           cn(
-            'fixed left-1/2 top-4 md:top-[calc(50%-250px)] z-50 w-[calc(100%-1rem)] max-w-screen-sm -translate-x-1/2 rounded-xl border bg-fd-popover text-fd-popover-foreground shadow-2xl shadow-black/50 overflow-hidden data-[closed]:animate-fd-dialog-out data-[open]:animate-fd-dialog-in',
+            'fixed left-1/2 top-4 md:top-[calc(50%-250px)] z-50 w-[calc(100%-1rem)] max-w-screen-sm -translate-x-1/2 rounded-xl border bg-fd-popover text-fd-popover-foreground shadow-2xl overflow-hidden data-closed:animate-fd-dialog-out data-open:animate-fd-dialog-in focus-visible:outline-none',
             '*:border-b *:has-[+:last-child[data-empty=true]]:border-b-0 *:data-[empty=true]:border-b-0 *:last:border-b-0',
             typeof className === 'function' ? className(s) : className,
           )
@@ -266,10 +368,12 @@ export function SearchDialogList({
     const viewport = element.firstElementChild;
     if (viewport) observer.observe(viewport);
 
-    window.addEventListener('keydown', onKey);
+    const content: Pick<Window, 'addEventListener' | 'removeEventListener'> =
+      document.getElementById('fd-search-dialog-content') ?? window;
+    content.addEventListener('keydown', onKey);
     return () => {
       observer.disconnect();
-      window.removeEventListener('keydown', onKey);
+      content.removeEventListener('keydown', onKey);
     };
   }, []);
 
@@ -292,7 +396,7 @@ export function SearchDialogList({
       <div
         className={cn('w-full flex flex-col overflow-y-auto max-h-[460px] p-1', !items && 'hidden')}
       >
-        <ListContext.Provider
+        <ListContext
           value={useMemo(
             () => ({
               active,
@@ -306,7 +410,7 @@ export function SearchDialogList({
           {items?.map((item) => (
             <Fragment key={item.id}>{Item({ item, onClick: () => onSelect(item) })}</Fragment>
           ))}
-        </ListContext.Provider>
+        </ListContext>
       </div>
     </div>
   );
@@ -316,10 +420,13 @@ export function SearchDialogListItem({
   item,
   className,
   children,
-  renderHighlights: render = renderHighlights,
+  renderMarkdown = (s) => <mdRenderer.Markdown components={mdComponents}>{s}</mdRenderer.Markdown>,
+  renderHighlights: _,
   ...props
 }: ComponentProps<'button'> & {
-  renderHighlights?: typeof renderHighlights;
+  renderMarkdown?: (v: string) => ReactNode;
+  /** @deprecated highlight blocks is now wrapped in `<mark />`, use `renderMarkdown` to handle instead. */
+  renderHighlights?: (blocks: HighlightedText<ReactNode>[]) => ReactNode;
   item: SearchItemType;
 }) {
   const { active: activeId, setActive } = useSearchList();
@@ -342,20 +449,21 @@ export function SearchDialogListItem({
         {item.type !== 'page' && (
           <div role="none" className="absolute start-3 inset-y-0 w-px bg-fd-border" />
         )}
-        <p
+        {item.type === 'heading' && (
+          <Hash className="absolute start-6 top-2.5 size-4 text-fd-muted-foreground" />
+        )}
+        <div
           className={cn(
-            'min-w-0 truncate',
-            item.type !== 'page' && 'ps-4',
+            'min-w-0',
+            item.type === 'text' && 'ps-4',
+            item.type === 'heading' && 'ps-8',
             item.type === 'page' || item.type === 'heading'
               ? 'font-medium'
               : 'text-fd-popover-foreground/80',
           )}
         >
-          {item.type === 'heading' && (
-            <Hash className="inline me-1 size-4 text-fd-muted-foreground" />
-          )}
-          {item.contentWithHighlights ? render(item.contentWithHighlights) : item.content}
-        </p>
+          {typeof item.content === 'string' ? renderMarkdown(item.content) : item.content}
+        </div>
       </>
     );
   }
@@ -377,7 +485,7 @@ export function SearchDialogListItem({
       )}
       aria-selected={active}
       className={cn(
-        'relative select-none px-2.5 py-2 text-start text-sm rounded-lg',
+        'relative select-none shrink-0 px-2.5 py-2 text-start text-sm overflow-hidden rounded-lg',
         active && 'bg-fd-accent text-fd-accent-foreground',
         className,
       )}
@@ -388,7 +496,6 @@ export function SearchDialogListItem({
     </button>
   );
 }
-
 export function SearchDialogIcon(props: ComponentProps<'svg'>) {
   const { isLoading } = useSearch();
 
@@ -420,22 +527,23 @@ const itemVariants = cva(
     },
   },
 );
-
 export function TagsList({ tag, onTagChange, allowClear = false, ...props }: TagsListProps) {
+  const onTagChangeCallback = useRef(onTagChange);
+  onTagChangeCallback.current = onTagChange;
   return (
     <div {...props} className={cn('flex items-center gap-1 flex-wrap', props.className)}>
-      <TagsListContext.Provider
+      <TagsListContext
         value={useMemo(
           () => ({
             value: tag,
-            onValueChange: onTagChange,
+            onValueChange: (v) => onTagChangeCallback.current(v),
             allowClear,
           }),
-          [allowClear, onTagChange, tag],
+          [allowClear, tag],
         )}
       >
         {props.children}
-      </TagsListContext.Provider>
+      </TagsListContext>
     </div>
   );
 }
@@ -455,9 +563,7 @@ export function TagsListItem({
       type="button"
       data-active={selected}
       className={cn(itemVariants({ active: selected, className }))}
-      onClick={() => {
-        onValueChange(selected && allowClear ? undefined : value);
-      }}
+      onClick={() => onValueChange(selected && allowClear ? undefined : value)}
       tabIndex={-1}
       {...props}
     >
@@ -466,22 +572,8 @@ export function TagsListItem({
   );
 }
 
-function renderHighlights(highlights: HighlightedText<ReactNode>[]): ReactNode {
-  return highlights.map((node, i) => {
-    if (node.styles?.highlight) {
-      return (
-        <span key={i} className="text-fd-primary underline">
-          {node.content}
-        </span>
-      );
-    }
-
-    return <Fragment key={i}>{node.content}</Fragment>;
-  });
-}
-
 export function useSearch() {
-  const ctx = use(Context);
+  const ctx = use(RootContext);
   if (!ctx) throw new Error('Missing <SearchDialog />');
   return ctx;
 }