+++ packages/openapi/src/ui/api-page.tsx
@@ -1,9 +1,8 @@
 import { Operation } from '@/ui/operation';
-import type { RenderContext } from '@/types';
+import type { HttpMethods, RenderContext } from '@/types';
 import { createMethod } from '@/utils/schema';
-import type { OpenAPIV3_1 } from 'openapi-types';
 import type { ProcessedDocument } from '@/utils/process-document';
-import { ApiProviderLazy } from './contexts/api.lazy';
+import { ApiProviderLazy, ServerProviderLazy } from './contexts/api.lazy';
 
 export interface ApiPageProps {
   document: Promise<ProcessedDocument> | string | ProcessedDocument;
@@ -23,7 +22,7 @@ export interface WebhookItem {
    * webhook name in `webhooks`
    */
   name: string;
-  method: OpenAPIV3_1.HttpMethods;
+  method: HttpMethods;
 }
 
 export interface OperationItem {
@@ -34,7 +33,7 @@ export interface OperationItem {
   /**
    * the HTTP method of operation
    */
-  method: OpenAPIV3_1.HttpMethods;
+  method: HttpMethods;
 }
 
 export async function APIPage({
@@ -113,14 +112,12 @@ export async function APIPage({
     },
     ctx,
   );
+  let servers = ctx.schema.dereferenced.servers;
+  if (!servers || servers.length === 0) servers = [{ url: '/' }];
 
   return (
-    <ApiProviderLazy
-      servers={ctx.servers}
-      shikiOptions={ctx.shikiOptions}
-      client={ctx.client ?? {}}
-    >
-      {content}
+    <ApiProviderLazy shikiOptions={ctx.shikiOptions} client={ctx.client ?? {}}>
+      <ServerProviderLazy servers={servers}>{content}</ServerProviderLazy>
     </ApiProviderLazy>
   );
 }